Refactor sidecar worker code.
